\name{AB}

\alias{AB}

\docType{data}

\title{Hypothetical AB data}

\description{Hypothetical data collected in an AB phase design with 27 measurement times.}

\usage{data(AB)}

\format{
  A data frame with 27 observations.
  \itemize{
    \item The first column contains the condition/phase labels ("A" and "B").
    \item The second column contains the obtained scores.
    \item The rows and columns are not labeled.
  }
}

\references{Bulte, I., & Onghena, P. (in press). When the Truth Hits You Between the Eyes: A Software Tool for the Visual Analysis of Single-Case Experimental Data. Manuscript accepted for publication in Methodology.}

\examples{data(AB)}

\keyword{datasets}
