\name{AB}

\alias{AB}

\docType{data}

\title{Hypothetical AB data}

\description{Hypothetical data collected in an AB phase design with 27 measurement times.}

\usage{data(AB)}

\format{
  A data frame with 27 observations.
  \itemize{
    \item The first column contains the condition/phase labels ("A" and "B").
    \item The second column contains the obtained scores.
    \item The rows and columns are not labeled.
  }
}

\references{Bulte, I., & Onghena, P. (2012). When the truth hits you between the eyes: A software tool for the visual analysis of single-case experimental data. Methodology, 8, 104-114.}

\examples{data(AB)}

\keyword{datasets}
