\name{SCVA-package}
\alias{SCVA-package}
\docType{package}
\title{
Single-Case Visual Analysis
}
\description{
A package to make graphical representations of single case data and to transform graphical displays back to raw data. The package also includes tools for visually analyzing single-case data, by displaying central location, variability and trend. 
}
\details{
\tabular{ll}{
Package: \tab SCVA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-01-13\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
Isis Bulte and Patrick Onghena

Maintainer: Marlies Vervloet <marlies.vervloet@ppw.kuleuven.be>
}

\keyword{ package }


